+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is simplified version of [Keep a Changelog](https://keepachangelog.co
 
 ### Additions
 - Add random library sort
+- Add the ability to save search queries
 
 ### Changes
 - Temporarily disable log file